Add some tests for combinatorial coverage.
